@@ -21,15 +21,15 @@
 	},
 	"license": "MIT",
 	"dependencies": {
-		"@angular/animations": "^9.1.13",
-		"@angular/cdk": "^9.2.4",
-		"@angular/common": "^9.1.13",
-		"@angular/compiler": "^9.1.13",
-		"@angular/core": "^9.1.13",
-		"@angular/forms": "^9.1.13",
-		"@angular/material": "^9.2.4",
-		"@angular/platform-browser": "^9.1.13",
-		"@angular/platform-browser-dynamic": "^9.1.13",
+		"@angular/animations": "^10.2.5",
+		"@angular/cdk": "^10.2.7",
+		"@angular/common": "^10.2.5",
+		"@angular/compiler": "^10.2.5",
+		"@angular/core": "^10.2.5",
+		"@angular/forms": "^10.2.5",
+		"@angular/material": "^10.2.7",
+		"@angular/platform-browser": "^10.2.5",
+		"@angular/platform-browser-dynamic": "^10.2.5",
 		"@firestone-hs/build-global-stats": "^1.0.4",
 		"@firestone-hs/categorize-deck": "^1.0.4",
 		"@firestone-hs/coliseum-components": "0.0.13",
@@ -49,7 +49,7 @@
 		"@types/reflect-metadata": "^0.1.0",
 		"angular-resize-event": "1.0.0",
 		"angular2-indexeddb": "1.2.3",
-		"angular2-notifications": "2.0.0",
+		"angular2-notifications": "9.0.0",
 		"aws-sdk": "2.458.0",
 		"chart.js": "^2.9.3",
 		"chartjs-plugin-datalabels": "^0.7.0",
@@ -59,12 +59,10 @@
 		"immutable": "3.8.2",
 		"jszip": "^3.2.2",
 		"lodash": "^4.17.20",
-		"ng-inline-svg": "8.4.1",
-		"ng-lz-string": "1.5.3",
+		"ng-inline-svg": "10.1.0",
 		"ng-select": "1.0.1",
 		"ng2-charts": "^2.3.0",
 		"ng2-fittext": "^1.2.8",
-		"ngx-logger": "3.3.11",
 		"ngx-perfect-scrollbar": "^8.0.0",
 		"node-vibrant": "^3.1.5",
 		"pako": "1.0.10",
@@ -78,10 +76,10 @@
 		"zone.js": "0.9.1"
 	},
 	"devDependencies": {
-		"@angular/compiler-cli": "9.1.13",
+		"@angular/compiler-cli": "10.2.5",
 		"@babel/preset-env": "^7.8.3",
 		"@calebmer/extract-text-webpack-plugin": "4.0.2",
-		"@ngtools/webpack": "^9.1.15",
+		"@ngtools/webpack": "^10.2.3",
 		"@overwolf/types": "^2.9.3",
 		"@sentry/cli": "^1.64.0",
 		"@sentry/webpack-plugin": "^1.15.0",
@@ -120,7 +118,7 @@
 		"to-string-loader": "^1.1.5",
 		"ts-jest": "^25.2.1",
 		"ts-loader": "^6.2.2",
-		"typescript": "^3.8.3",
+		"typescript": "^3.9.4",
 		"webpack": "4.41.1",
 		"webpack-bundle-analyzer": "3.4.1",
 		"webpack-cli": "^3.3.11",
