@@ -1,7 +1,6 @@
 import { ModuleWithProviders, NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { AllCardsService } from '@firestone-hs/replay-parser';
-import { LoggerModule, NgxLoggerLevel } from 'ngx-logger';
 import { AchievementHistoryStorageService } from '../../services/achievement/achievement-history-storage.service';
 import { AchievementsRepository } from '../../services/achievement/achievements-repository.service';
 import { ChallengeBuilderService } from '../../services/achievement/achievements/challenges/challenge-builder.service';
@@ -28,13 +27,13 @@ import { PreferencesService } from '../../services/preferences.service';
 import { S3FileUploadService } from '../../services/s3-file-upload.service';
 
 @NgModule({
-	imports: [BrowserModule, LoggerModule.forRoot({ level: NgxLoggerLevel.DEBUG })],
+	imports: [BrowserModule],
 	declarations: [],
 	entryComponents: [],
 	exports: [],
 })
 export class SharedServicesModule {
-	static forRoot(): ModuleWithProviders {
+	static forRoot(): ModuleWithProviders<any> {
 		return {
 			ngModule: SharedServicesModule,
 			providers: [
